Modify code-of-conduct file
In the assignment 4, we got a feedback about what we did and we revising this feedback for the assignment 7.

We add the reflections about the LLM suggestions for "How do we ensure knowledge transfer among our team members?",
where we insert what we will use, how and why for each of the suggestion, additionally we compared our answer with the LLM answer, to see the different prefaces bewteen them.

Secondly, we add the sub-information and fullfil description about the question number 6 "How do you plan to ensure responsible use of LLMs in your project and how do you transparently and traceably document the use of LLMs?",
where we provide more description about how each of the team members will be responsible about the LLM transparently, and that by make each of the additions as mimics points how and where and why we should provide the LLM modification on the specific change and why we are going to use that suggetions.

Rersolves: #1