Avoid unnecessery build
Modified the GitLab CI/CD YAML file to introduce a caching mechanism for Docker image builds. This caching mechanism pulls the cached image based on the commit reference slug before building the Docker image. If the cached image is not available, the regular build process continues as usual. This optimization helps avoid unnecessary builds for known successful commits, improving the efficiency of the CI/CD pipeline.